third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-image-in-iframe-expected.txt

CONSOLE WARNING: line 1: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-image.html' was loaded over HTTPS, but displayed insecure content from 'http://127.0.0.1:8080/security/resources/compass.jpg': this content should also be loaded over HTTPS.

This test loads a secure iframe that loads an insecure image. We should get a mixed content callback becase the frame is HTTPS and the image is not.
